refactor KVServiceManagers: Move IllegalParameterCountException to a shared module
