+++ kvision-modules/kvision-common-remote/src/jvmMain/kotlin/pl/treksoft/kvision/remote/IllegalParameterCountException.kt
@@ -1,0 +1,17 @@
+package pl.treksoft.kvision.remote
+
+class IllegalParameterCountException(val actualParameterCount: Int, val expectedParamterCount: Int) :
+    RuntimeException() {
+    override val message: String
+        get() = "Expected <$expectedParamterCount> parameters, but got <$actualParameterCount> parameters"
+}
+
+fun requireParameterCountEqualTo(actualParameterCount: Int, expectedParamterCount: Int) {
+    if (actualParameterCount != expectedParamterCount) {
+        throw IllegalParameterCountException(actualParameterCount, expectedParamterCount)
+    }
+}
+
+fun requireParameterCountEqualTo(params: Collection<*>, expectedParameterCount: Int) =
+    requireParameterCountEqualTo(params.size, expectedParameterCount)
+
