+++ kvision-modules/kvision-server-ktor/src/jvmMain/kotlin/pl/treksoft/kvision/remote/KVServiceManager.kt
@@ -382,21 +382,6 @@ actual open class KVServiceManager<T : Any> actual constructor(val serviceClass:
     }
 }
 
-class IllegalParameterCountException(val actualParameterCount: Int, val expectedParamterCount: Int) :
-    RuntimeException() {
-    override val message: String
-        get() = "Expected <$expectedParamterCount> parameters, but got <$actualParameterCount> parameters"
-}
-
-fun requireParameterCountEqualTo(actualParameterCount: Int, expectedParamterCount: Int) {
-    if (actualParameterCount != expectedParamterCount) {
-        throw IllegalParameterCountException(actualParameterCount, expectedParamterCount)
-    }
-}
-
-fun requireParameterCountEqualTo(params: Collection<*>, expectedParameterCount: Int) =
-    requireParameterCountEqualTo(params.size, expectedParameterCount)
-
 /**
  * A function to generate routes based on definitions from the service manager.
  */
