+++ kvision-modules/kvision-server-spring-boot/src/jvmMain/kotlin/pl/treksoft/kvision/remote/KVServiceManager.kt
@@ -394,18 +394,3 @@ actual open class KVServiceManager<T : Any> actual constructor(val serviceClass:
         } ?: null as T
     }
 }
-
-class IllegalParameterCountException(val actualParameterCount: Int, val expectedParamterCount: Int) :
-    RuntimeException() {
-    override val message: String
-        get() = "Expected <$expectedParamterCount> parameters, but got <$actualParameterCount> parameters"
-}
-
-fun requireParameterCountEqualTo(actualParameterCount: Int, expectedParamterCount: Int) {
-    if (actualParameterCount != expectedParamterCount) {
-        throw IllegalParameterCountException(actualParameterCount, expectedParamterCount)
-    }
-}
-
-fun requireParameterCountEqualTo(params: Collection<*>, expectedParameterCount: Int) =
-    requireParameterCountEqualTo(params.size, expectedParameterCount)
